Change the tethering upstream priorities.
We had mobile_dun before wimax, so if wimax ever dropped out we'd switch
to mobile_dun and never give it up.  Reverse the order so that mobile_dun
is the connection of last resort.  If wimax comes back we should now switch
back to using it for the tether upstream connection.

Change-Id: If8000e36f6137aeec74c8cdcc8a636933f406b46
bug:6045300