Add "config_cellBroadcastAppLinks" to Crespo 4G overlay.
Set the new "config_cellBroadcastAppLinks" item to true for
Sprint Nexus S. This will add a new menu item to the MMS app,
for direct access to the list of received Cell Broadcasts, and
a new item in Settings to jump to Cell Broadcast settings.
This also requires that "CellBroadcastReceiver" is added to the
list of PRODUCT_PACKAGES in a makefile for this device config.

Bug: 6709985
Change-Id: I02dc7d8cf6982a2f4d148cc856e108789421b362